Optimise index reload detection
Generate an identity file which broadcasts the latest unique ID since an
update. If the identity changes (i.e. is detected by the parent page)
then the index is reloaded and reparsed.

This avoids reloading and reparsing a potentially large index unless
there is data to update, so the reload time for the identity page can be
quick without loading the CPU with parsing/decoding/re-encoding/etc.