Address performance issues by avoiding unnecessary processing
We do not need to constantly spam the Index iframe with HELO messages
(which will cause the iframe to re-parse its content) - we can just send
messages between the time that we reload the Index (due to the Identity
iframe updating) and when the index has successfully updated.

This massively reduces CPU overhead as the data that needs to be parsed
can end up being pretty significant.